% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_f_k}
\alias{get_f_k}
\title{Get the f(k) vector}
\usage{
get_f_k(obj)
}
\arguments{
\item{obj}{the output of kselection function}
}
\value{
the vector of f(k) function
}
\description{
Get the f(k) vector
}
\examples{
# Create a data set with two clusters
dat <- matrix(c(rnorm(100, 2, .1), rnorm(100, 3, .1),
                rnorm(100, -2, .1), rnorm(100, -3, .1)), 200, 2)

# Get the f(k) vector
sol <- kselection(dat)
f_k <- get_f_k(sol)
}
\seealso{
\code{\link{num_clusters}},
}

