% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{kselection-package}
\alias{kselection-package}
\title{kselection: Selection of k in k-means clustering}
\description{
Selection of k in k-means clustering based on Pham et al. paper ``Selection
of k in k-means clustering''
}
\details{
\tabular{ll}{
  Package: \tab kselection \cr
  Version: \tab 0.1.0 \cr
  License: \tab GPL-3 \cr
}
}
\author{
Daniel Rodriguez Perez \email{daniel.rodriguez.perez@gmail.com}
}
\references{
D T Pham, S S Dimov, and C D Nguyen "Selection of k in k-means clustering",
Mechanical Engineering Science, 2004, pp. 103-119.
}

