% Generated by roxygen2 (4.0.2): do not edit by hand
\name{kselection}
\alias{kselection}
\title{Selection of K in K-means Clustering}
\usage{
kselection(x, max_centers = 15, k_threshold = 0.85, iter.max = 200,
  nstart = 100, trace = FALSE)
}
\arguments{
\item{x}{numeric matrix of data, or an object that can be coerced to such a
matrix.}

\item{max_centers}{maximum number of clusters for evaluation.}

\item{k_threshold}{cluster selection threshold for f(k) value.}

\item{iter.max}{the maximum number of iterations allowed.}

\item{nstart}{if centers is a number, how many random sets should be chosen?}

\item{trace}{show a progress bar}
}
\value{
an array with f(k) metric
}
\description{
Selection of k in k-means clustering based on Pham et al. paper
}
\author{
Daniel Rodriguez Perez
}
\references{
D T Pham, S S Dimov, and C D Nguyen "Selection of k in k-means clustering",
Mechanical Engineering Science, 2004, pp. 103-119.
}

