% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pham.R
\name{set_k_threshold}
\alias{set_k_threshold}
\title{Set the \code{k_threshold}}
\usage{
set_k_threshold(obj, k_threshold)
}
\arguments{
\item{obj}{the output of \code{kselection} function.}

\item{k_threshold}{maximum value of \eqn{f(K)} from which can not be
       considered the existence of more than one cluster in the data set.}
}
\value{
the output of kselection function with new \code{k_threshold}.
}
\description{
Set the maximum value of \eqn{f(K)} from which can not be considered the
existence of more than one cluster.
}
\author{
Daniel Rodriguez
}
\seealso{
\code{\link{get_k_threshold}}
}

