% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_k_threshold}
\alias{get_k_threshold}
\title{Get the \code{k_threshold}}
\usage{
get_k_threshold(obj)
}
\arguments{
\item{obj}{the output of \code{kselection} function.}
}
\value{
the \code{k_threshold} value.
}
\description{
Get the maximum value of \eqn{f(K)} from which can not be considered the
existence of more than one cluster.
}
\author{
Daniel Rodriguez Perez
}
\seealso{
\code{\link{set_k_threshold}}
}

