% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kselection.R
\docType{package}
\name{kselection-package}
\alias{kselection-package}
\title{Selection of K in K-Means Clustering}
\description{
Selection of k in k-means clustering based on Pham et al. paper ``Selection
of k in k-means clustering''
}
\details{
This package implements the method for selecting the number of clusters for
the algorithm K-means introduced in the publication of Pham, Dimov and
Nguyen of 2004.

\tabular{ll}{
  Package: \tab kselection \cr
  Version: \tab 0.2.0 \cr
  License: \tab GPL-3 \cr
}
}
\author{
Daniel Rodriguez \email{daniel.rodriguez.perez@gmail.com}
}
\references{
D T Pham, S S Dimov, and C D Nguyen, "Selection of k in k-means clustering",
Mechanical Engineering Science, 2004, pp. 103-119.
}

