% Generated by roxygen2 (4.0.2): do not edit by hand
\name{kselection}
\alias{kselection}
\title{Selection of K in K-means Clustering}
\usage{
kselection(x, max_centers = 15, k_threshold = 0.85, iter.max = 200,
  nstart = 100, trace = FALSE)
}
\arguments{
\item{x}{numeric matrix of data, or an object that can be coerced to such a
matrix.}

\item{max_centers}{maximum number of clusters for evaluation.}

\item{k_threshold}{cluster selection threshold for f(k) value.}

\item{iter.max}{the maximum number of iterations allowed.}

\item{nstart}{if centers is a number, how many random sets should be chosen?}

\item{trace}{show a progress bar}
}
\value{
an object with the f(k) results
}
\description{
Selection of k in k-means clustering based on Pham et al. paper
}
\details{
This function implements the method for selecting the number of clusters for
the algorithm K-means introduced in the publication of Pham, Dimov and
Nguyen of 2004.

The method introduce a function f(k) to evaluate the quality of the
resulting clustering. The values of k which yield small values of f(k) can
be considered to produce well-defined clusters.
}
\examples{
# Create a data set with two clusters
dat <- matrix(c(rnorm(100, 2, .1), rnorm(100, 3, .1),
                rnorm(100, -2, .1), rnorm(100, -3, .1)), 200, 2)

# Ejecute the method
sol <- kselection(dat)

# Get the results
k   <- num_clusters(sol) # optimal number of clustes
f_k <- get_f_k(sol)      # the f(k) vector
}
\author{
Daniel Rodriguez Perez
}
\references{
D T Pham, S S Dimov, and C D Nguyen "Selection of k in k-means clustering",
Mechanical Engineering Science, 2004, pp. 103-119.
}

