% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pham.R
\name{num_clusters}
\alias{num_clusters}
\title{Get the number of clusters.}
\usage{
num_clusters(obj)
}
\arguments{
\item{obj}{the output of kselection function.}
}
\value{
the number of clusters proposed.
}
\description{
The optimal number of clusters proposed by the method.
}
\examples{
# Create a data set with two clusters
dat <- matrix(c(rnorm(100, 2, .1), rnorm(100, 3, .1),
                rnorm(100, -2, .1), rnorm(100, -3, .1)), 200, 2)

# Get the optimal number of clustes
sol <- kselection(dat)
k   <- num_clusters(sol)
}
\author{
Daniel Rodriguez
}
\seealso{
\code{\link{num_clusters_all}}, \code{\link{get_f_k}}
}

